+++ src/constants/box-shadow.js
@@ -1,0 +1,18 @@
+// @flow
+
+import { css } from 'styled-components';
+
+const BS_SMALL = css`
+  box-shadow: 15px 18px 15px -15px rgba(0, 0, 0, 0.1);
+`;
+
+const BS_LARGE = css`
+  box-shadow: 32px 64px 40px -30px rgba(0, 0, 0, 0.1);
+`;
+
+const BOX_SHADOW = {
+  BS_SMALL,
+  BS_LARGE,
+};
+
+export default BOX_SHADOW;