+++ src/constants/letter-spacing.js
@@ -1,0 +1,18 @@
+// @flow
+
+import { css } from 'styled-components';
+
+const LS_1 = css`
+  letter-spacing: 1px;
+`;
+
+const LS_2_5 = css`
+  letter-spacing: 2.5px;
+`;
+
+const LETTER_SPACING = {
+  LS_1,
+  LS_2_5,
+};
+
+export default LETTER_SPACING;