+++ src/index.js
@@ -2,10 +2,12 @@
 
 export { default as BORDER_RADIUS } from './constants/border-radius';
 export { default as BORDER_WIDTH } from './constants/border-width';
+export { default as BOX_SHADOW } from './constants/box-shadow';
 export { default as COLORS } from './constants/colors';
 export { default as TYPE_SCALE } from './constants/type-scale';
 export { default as FONT_WEIGHT } from './constants/font-weight';
 export { default as LINE_HEIGHT } from './constants/line-height';
+export { default as LETTER_SPACING } from './constants/letter-spacing';
 export { default as MAX_WIDTH } from './constants/max-width';
 export { default as MEDIA } from './constants/media-queries';
 export { default as TRANSITIONS } from './constants/transitions';
@@ -19,3 +21,4 @@ export { default as Button } from './button';
 export { default as Modal } from './modal';
 export { default as ToolTip } from './tooltip';
 export { default as Spinner } from './spinner';
+export { default as Blockquote } from './blockquote';